+++ examples/resources/aerospike_sindex/import.sh
@@ -1,0 +1,1 @@
+terraform import aerospike_sindex.jobs aerospike/shuttlex_jobs/shuttlex_jobs-set-idx