+++ internal/provider/provider.go
@@ -244,6 +244,7 @@ func (p *AerospikeProvider) Resources(ctx context.Context) []func() resource.Res
 		NewAerospikeNamespaceConfig,
 		NewAerospikeServiceConfig,
 		NewAerospikeXDRDCConfig,
+		NewAerospikeSindex,
 	}
 }